refactor(slack): split adapter into focused modules
