+++ src/channels/slack-adapter-events.test.ts
@@ -1,0 +1,514 @@
+import { expect, mock, test } from "bun:test";
+import {
+  createSlackAdapter,
+  FakeSlackApp,
+  installSlackAdapterTestHooks,
+  resolveSlackInboundAttachmentsMock,
+  slackAccountDefaults,
+} from "./slack-adapter.test-harness";
+
+installSlackAdapterTestHooks();
+
+test("slack adapter dedupes threaded mentions delivered through message and app_mention", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  const onMessage = mock(async () => {});
+  adapter.onMessage = onMessage;
+
+  await adapter.start();
+  const app = FakeSlackApp.instances[0];
+  const messageHandler = app?.messageHandler;
+  const mentionHandler = app?.eventHandlers.get("app_mention");
+  if (!messageHandler || !mentionHandler) {
+    throw new Error("Expected Slack message and mention handlers");
+  }
+
+  await messageHandler({
+    message: {
+      channel: "C123",
+      user: "U123",
+      text: "<@U0AS42PTEAX> following up in thread",
+      ts: "1712800000.000100",
+      thread_ts: "1712790000.000050",
+    },
+  });
+
+  await mentionHandler({
+    event: {
+      channel: "C123",
+      user: "U123",
+      text: "<@U0AS42PTEAX> following up in thread",
+      ts: "1712800000.000100",
+      thread_ts: "1712790000.000050",
+    },
+  });
+
+  expect(onMessage).toHaveBeenCalledTimes(1);
+  expect(onMessage).toHaveBeenCalledWith(
+    expect.objectContaining({
+      channel: "slack",
+      chatId: "C123",
+      senderId: "U123",
+      text: "following up in thread",
+      messageId: "1712800000.000100",
+      threadId: "1712790000.000050",
+      chatType: "channel",
+      isMention: true,
+    }),
+  );
+});
+
+test("slack adapter dedupes threaded mentions when app_mention arrives first", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  const onMessage = mock(async () => {});
+  adapter.onMessage = onMessage;
+
+  await adapter.start();
+  const app = FakeSlackApp.instances[0];
+  const messageHandler = app?.messageHandler;
+  const mentionHandler = app?.eventHandlers.get("app_mention");
+  if (!messageHandler || !mentionHandler) {
+    throw new Error("Expected Slack message and mention handlers");
+  }
+
+  await mentionHandler({
+    event: {
+      channel: "C123",
+      user: "U123",
+      text: "<@U0AS42PTEAX> still there?",
+      ts: "1712800000.000101",
+      thread_ts: "1712790000.000050",
+    },
+  });
+
+  await messageHandler({
+    message: {
+      channel: "C123",
+      user: "U123",
+      text: "<@U0AS42PTEAX> still there?",
+      ts: "1712800000.000101",
+      thread_ts: "1712790000.000050",
+    },
+  });
+
+  expect(onMessage).toHaveBeenCalledTimes(1);
+  expect(onMessage).toHaveBeenCalledWith(
+    expect.objectContaining({
+      channel: "slack",
+      chatId: "C123",
+      senderId: "U123",
+      text: "still there?",
+      messageId: "1712800000.000101",
+      threadId: "1712790000.000050",
+      chatType: "channel",
+      isMention: true,
+    }),
+  );
+});
+
+test("slack adapter allows file_share subtype messages through", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  resolveSlackInboundAttachmentsMock.mockResolvedValueOnce([
+    {
+      id: "F123",
+      name: "screenshot.png",
+      mimeType: "image/png",
+      kind: "image",
+      localPath: "/tmp/screenshot.png",
+    },
+  ]);
+
+  const onMessage = mock(async () => {});
+  adapter.onMessage = onMessage;
+
+  await adapter.start();
+  const app = FakeSlackApp.instances[0];
+  const handler = app?.messageHandler;
+  if (!handler) {
+    throw new Error("Expected Slack message handler");
+  }
+
+  await handler({
+    message: {
+      channel: "C123",
+      user: "U123",
+      text: "",
+      ts: "1712800000.000100",
+      thread_ts: "1712790000.000050",
+      subtype: "file_share",
+      files: [{ id: "F123", name: "screenshot.png" }],
+    },
+  });
+
+  expect(onMessage).toHaveBeenCalledWith(
+    expect.objectContaining({
+      chatId: "C123",
+      threadId: "1712790000.000050",
+      attachments: [
+        expect.objectContaining({
+          id: "F123",
+          name: "screenshot.png",
+        }),
+      ],
+    }),
+  );
+});
+
+test("slack adapter passes transcription opt-in to message and app_mention media resolution", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+    transcribeVoice: true,
+  });
+
+  adapter.onMessage = mock(async () => {});
+
+  await adapter.start();
+  const app = FakeSlackApp.instances[0];
+  const messageHandler = app?.messageHandler;
+  const mentionHandler = app?.eventHandlers.get("app_mention");
+  if (!messageHandler || !mentionHandler) {
+    throw new Error("Expected Slack handlers");
+  }
+
+  await messageHandler({
+    message: {
+      channel: "D123",
+      user: "U123",
+      text: "voice note",
+      ts: "1712800000.000100",
+    },
+  });
+  await mentionHandler({
+    event: {
+      channel: "C123",
+      user: "U123",
+      text: "<@U0AS42PTEAX> voice note",
+      ts: "1712800000.000101",
+    },
+  });
+
+  expect(resolveSlackInboundAttachmentsMock).toHaveBeenCalledTimes(2);
+  expect(resolveSlackInboundAttachmentsMock).toHaveBeenNthCalledWith(
+    1,
+    expect.objectContaining({
+      accountId: "slack-test-account",
+      token: "xoxb-test-token-1234567890",
+      rawEvent: expect.objectContaining({ channel: "D123" }),
+      transcribeVoice: true,
+    }),
+  );
+  expect(resolveSlackInboundAttachmentsMock).toHaveBeenNthCalledWith(
+    2,
+    expect.objectContaining({
+      accountId: "slack-test-account",
+      token: "xoxb-test-token-1234567890",
+      rawEvent: expect.objectContaining({ channel: "C123" }),
+      transcribeVoice: true,
+    }),
+  );
+});
+
+test("slack adapter allows thread_broadcast subtype replies through", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  const onMessage = mock(async () => {});
+  adapter.onMessage = onMessage;
+
+  await adapter.start();
+  const app = FakeSlackApp.instances[0];
+  const handler = app?.messageHandler;
+  if (!handler) {
+    throw new Error("Expected Slack message handler");
+  }
+
+  await handler({
+    message: {
+      channel: "C123",
+      user: "U123",
+      text: "broadcasting this reply",
+      ts: "1712800000.000101",
+      thread_ts: "1712790000.000050",
+      subtype: "thread_broadcast",
+    },
+  });
+
+  expect(onMessage).toHaveBeenCalledWith(
+    expect.objectContaining({
+      channel: "slack",
+      chatId: "C123",
+      senderId: "U123",
+      text: "broadcasting this reply",
+      messageId: "1712800000.000101",
+      threadId: "1712790000.000050",
+      chatType: "channel",
+      isMention: false,
+    }),
+  );
+});
+
+test("slack adapter ignores hidden bookkeeping thread wrapper events", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  const onMessage = mock(async () => {});
+  adapter.onMessage = onMessage;
+
+  await adapter.start();
+  const app = FakeSlackApp.instances[0];
+  const handler = app?.messageHandler;
+  if (!handler) {
+    throw new Error("Expected Slack message handler");
+  }
+
+  await handler({
+    message: {
+      channel: "C123",
+      ts: "1712800000.000102",
+      hidden: true,
+      subtype: "message_replied",
+      message: {
+        type: "message",
+        user: "U123",
+        text: "Original thread root",
+        thread_ts: "1712790000.000050",
+        ts: "1712790000.000050",
+      },
+    },
+  });
+
+  expect(onMessage).not.toHaveBeenCalled();
+});
+
+test("slack adapter ignores live bot_message events as runnable input", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  const onMessage = mock(async () => {});
+  adapter.onMessage = onMessage;
+
+  await adapter.start();
+  const app = FakeSlackApp.instances[0];
+  const handler = app?.messageHandler;
+  if (!handler) {
+    throw new Error("Expected Slack message handler");
+  }
+
+  await handler({
+    message: {
+      channel: "C123",
+      bot_id: "BDEPLOY",
+      text: "deployment completed",
+      ts: "1712800000.000103",
+      thread_ts: "1712790000.000050",
+      subtype: "bot_message",
+    },
+  });
+
+  expect(onMessage).not.toHaveBeenCalled();
+  expect(resolveSlackInboundAttachmentsMock).not.toHaveBeenCalled();
+});
+
+test("slack adapter forwards reaction events into the routed Slack thread", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  const onMessage = mock(async () => {});
+  adapter.onMessage = onMessage;
+
+  await adapter.start();
+  const app = FakeSlackApp.instances[0];
+  const messageHandler = app?.messageHandler;
+  const reactionHandler = app?.eventHandlers.get("reaction_added");
+  if (!messageHandler || !reactionHandler) {
+    throw new Error("Expected Slack message and reaction handlers");
+  }
+
+  await messageHandler({
+    message: {
+      channel: "C123",
+      user: "U123",
+      text: "following up in thread",
+      ts: "1712800000.000100",
+      thread_ts: "1712790000.000050",
+    },
+  });
+
+  onMessage.mockClear();
+
+  await reactionHandler({
+    event: {
+      user: "U555",
+      item_user: "U123",
+      reaction: "eyes",
+      event_ts: "1712800001.000200",
+      item: {
+        type: "message",
+        channel: "C123",
+        ts: "1712800000.000100",
+      },
+    },
+  });
+
+  expect(onMessage).toHaveBeenCalledWith(
+    expect.objectContaining({
+      chatId: "C123",
+      chatType: "channel",
+      threadId: "1712790000.000050",
+      text: "Slack reaction added: :eyes:",
+      reaction: {
+        action: "added",
+        emoji: "eyes",
+        targetMessageId: "1712800000.000100",
+        targetSenderId: "U123",
+      },
+    }),
+  );
+});
+
+test("slack adapter ignores reactions authored by its own bot user", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  const onMessage = mock(async () => {});
+  adapter.onMessage = onMessage;
+
+  await adapter.start();
+  const app = FakeSlackApp.instances[0];
+  const reactionAddedHandler = app?.eventHandlers.get("reaction_added");
+  const reactionRemovedHandler = app?.eventHandlers.get("reaction_removed");
+  if (!reactionAddedHandler || !reactionRemovedHandler) {
+    throw new Error("Expected Slack reaction handlers");
+  }
+
+  const event = {
+    user: "U0AS42PTEAX",
+    item_user: "U123",
+    reaction: "x",
+    event_ts: "1712800001.000200",
+    item: {
+      type: "message",
+      channel: "C123",
+      ts: "1712800000.000100",
+    },
+  };
+
+  await reactionAddedHandler({ event });
+  await reactionRemovedHandler({ event });
+
+  expect(onMessage).not.toHaveBeenCalled();
+});
+
+test("slack adapter preserves non-leading user mentions in app mention text", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  const onMessage = mock(async () => {});
+  adapter.onMessage = onMessage;
+
+  await adapter.start();
+  const app = FakeSlackApp.instances[0];
+  const handler = app?.eventHandlers.get("app_mention");
+  if (!handler) {
+    throw new Error("Expected app_mention handler");
+  }
+
+  await handler({
+    event: {
+      channel: "C123",
+      user: "U123",
+      text: "<@U999> ask <@U555> for help",
+      ts: "1712800000.000100",
+    },
+  });
+
+  expect(onMessage).toHaveBeenCalledWith(
+    expect.objectContaining({
+      text: "ask <@U555> for help",
+    }),
+  );
+});
